+++ Database/Models/GeometricParametersModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace Database.Models;
 
 public record GeometricParametersModel(
-    int Height,
-    int Length,
-    int Width);
+    double Height,
+    double Length,
+    double Width);
